Update Flowchart.js
instead of using a separate function, just put the functionality into the default function. in order to use a state variable to update the page, which seemed to be the simplest way, everything needed to be contained within one function. I love abstraction, but keeping it as separate functions would have complicated everyone's lives, and the way it's set up now isn't the most complicated thing in the world so I think it's fine.