Update Flowchart.js
Updates to comments, but also wanted to denote in a commit that, at current, any of the questions that would lead to another page currently just take you back to the landing page. when the whole app is integrated together we can easily edit hierarchy.json to allow the questions to link to other pages as needed.